requirements.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.0.9
django-3-jet==1.0.8
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
